% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fulldash.R
\name{rRomaDash}
\alias{rRomaDash}
\title{Launch the rRoma dashboard}
\usage{
rRomaDash(RomaData = NULL, ExpMat = NULL, Groups = NULL,
  Interactive = FALSE)
}
\arguments{
\item{Interactive}{boolean, scalar. Should interactivity (via Plotly) be enabled?}
}
\description{
Launch the rRoma dashboard
}
