% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SubSetRoma}
\alias{SubSetRoma}
\title{Subset an rRomaDash save file}
\usage{
SubSetRoma(SourceFile, TargetFile, KeyWord)
}
\arguments{
\item{SourceFile}{the source RDS file}

\item{TargetFile}{the target RDS file}

\item{KeyWord}{the keywork to use to subset the original save file}
}
\description{
Subset an rRomaDash save file
}
